# Re-implement LatestMap as a sorted slice for better performance