--- a/report/latest_map_internal_test.go
+++ b/report/latest_map_internal_test.go
@@ -72,7 +72,7 @@ func TestLatestMapDeepEquals(t *testing.T) {
 
 func nilStringLatestMap() StringLatestMap {
 	m := MakeStringLatestMap()
-	m.Map = nil
+	m.entries = nil
 	return m
 }