Part of fixing #2866 

## Benchmarks
### Before
```
$ go test -tags netgo -bench=LatestMap
BenchmarkLatestMapMerge-2    	    3000	    403395 ns/op	  559041 B/op	    4993 allocs/op
BenchmarkLatestMapEncode-2   	    5000	    238557 ns/op	   98768 B/op	    1014 allocs/op
BenchmarkLatestMapDecode-2   	    5000	    317857 ns/op	  164944 B/op	    4004 allocs/op
```
### After
```
$ go test -tags netgo -bench=LatestMap
BenchmarkLatestMapMerge-2    	   30000	     53852 ns/op	  131072 B/op	       1 allocs/op
BenchmarkLatestMapEncode-2   	   10000	    216475 ns/op	   98768 B/op	    1014 allocs/op
BenchmarkLatestMapDecode-2   	    5000	    287478 ns/op	   70492 B/op	    2005 allocs/op
```