@bboreham 2017-09-28
Currently this is implemented with append-like semantics on add at the end, i.e. if I have

    x := {"a":"1", "b":"2"}
    y := x.Set("c","3")

then x and y may share the same storage. Almost everything will work as expected, but if they are sharing the storage and you then do:

    z := x.Set("d","4")

then this will change y too.

I guess I should change that.