@rade 2017-10-10
iirc rendering is quite heavy on LatestMap lookups.

I'd like to see some comparative results for pumping real reports through various stages of processing. I might have a go at that myself, since I did quite a lot of that in the past using
```
prog/scope --mode=app --weave=false --app.collector=file:/some-report.json
```
and then timing something like
```
time curl -so /dev/null http://localhost:4040/api/topology/hosts
```

extras/copyreport is also handy for benchmarking.