@rade 2017-10-11
This PR makes no difference to rendering performance. Which I suppose is a good thing.

I grabbed a current Weave Cloud production report from the UI and ran
```
prog/scope --mode=app --weave=false --app.collector=file:/home/matthias/tmp/prod-on-prod-report.json
```
and then
```
time curl -so /dev/null http://localhost:4040/api/topology/hosts
```
I repeated this for a few other topology types, taking care to restart scope every time since otherwise rendering hits the internal gcache.

|                  | old | new |
| ------------------ | ---- | ----- |
| hosts            | 1.3 | 1.3 |
| kube-controllers | 1.0 | 1.0 |
| pods             | 1.0 | 1.0 |
| containers       | 1.3 | 1.3 |
| processes        | 0.3 | 0.3 |

where old=master and new=this-branch.