@rade 2017-10-11
I don't see much difference when copying a report, perhaps because it's dominated by writing:
```
extras/copyreport/copyreport ~/tmp/prod-on-prod-report.json ~/tmp/prod-on-prod-report.msgpack.gz
time extras/copyreport/copyreport ~/tmp/prod-on-prod-report.msgpack.gz /tmp/report.msgpack.gz
```
old=2.6s, new=2.5s

I then applied
```diff
index 434edcd8..5bc5b135 100644
--- a/app/collector.go
+++ b/app/collector.go
@@ -248,6 +248,7 @@ func NewFileCollector(path string, window time.Duration) (Collector, error) {
                timestamps []time.Time
                reports    []report.Report
        )
+       start := time.Now()
        allTimestamped := true
        if err := filepath.Walk(path,
                func(p string, info os.FileInfo, err error) error {
@@ -272,12 +273,12 @@ func NewFileCollector(path string, window time.Duration) (Collector, error) {
                }); err != nil {
                return nil, err
        }
-       if len(reports) > 1 && allTimestamped {
-               collector := NewCollector(window)
-               go replay(collector, timestamps, reports)
-               return collector, nil
-       }
-       return StaticCollector(NewSmartMerger().Merge(reports).Upgrade()), nil
+       fmt.Println("read in", time.Since(start))
+       m := NewSmartMerger()
+       start = time.Now()
+       r := m.Merge(reports)
+       fmt.Println("merged in", time.Since(start))
+       return StaticCollector(r.Upgrade()), nil
 }
 
 func timestampFromFilepath(path string) (time.Time, error) {
```
and ran
```
prog/scope --mode=app --weave=false --app.collector=file:/home/matthias/tmp/reports2
```
where the specified directory contains 30s worth of Weave Cloud prod reports from a few months ago, in msgpack.gz format. I did five runs and averaged the results:

| | old | new |
| -- | -- | -- |
| read | 14.0 | 13.0 |
| merge | 6.0 | 4.9 |

So that's a reasonable improvement in read performance and a significant improvement in merge performance.

Let's get this PR merged!