@bboreham 2017-10-11
I can recommend the Go benchmark support for this kind of experiment - you call `ResetTimer()` after initialization and it will increase the loop count until a certain time has been exceeded.